chore: correct python folder spelling
