+++ Python/briefpython.py
